IoT/flask_api/requirements.txt
flask
SpeechRecognition
pyttsx3
PyAudio
pyaudio
pyserial
